Remove old upload worker and associated structs
